+++ web/admin/app.js
@@ -12,6 +12,7 @@ import {
 } from "./js/ui-shell.js";
 import { initOverlayPreview, updateOverlayPreviewScale } from "./js/overlay-preview.js";
 import { initOBSSetup } from "./js/obs-setup.js";
+import { initConnectionsTabs } from "./js/connections.js";
 import { initSettingsDialogs } from "./js/dialogs.js";
 import { initAboutDialog } from "./js/about.js";
 import { initMessageSoundControls } from "./js/sound.js";
@@ -74,6 +75,7 @@ handleOAuthQuery();
 initSidebarToggle();
 initOverlayPreview();
 initOBSSetup();
+initConnectionsTabs();
 initSettingsDialogs();
 initAboutDialog();
 initMessageSoundControls();